+++ src/main/java/com/example/marketgospring/repository/GoodsRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface GoodsRepository extends JpaRepository <Goods, Long> {
+public interface GoodsRepository extends JpaRepository <Goods, Integer> {
     //Optional<Goods> findByStoreId(Long storeId);
 
     //Optional<Goods> findByMarketId(Long marketId);
