+++ src/main/java/com/example/marketgospring/repository/S3FileRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface S3FileRepository extends JpaRepository<S3File, Long> {
+public interface S3FileRepository extends JpaRepository<S3File, Integer> {
 }
